fix: room couese list error in YanChao campus
Fixes the problem where querying classroom schedules for Yan-Chao campus was not functioning correctly. This issue was caused by the inability to retrieve data for classroom MA320 across the entire university. To address this, a condition has been implemented that automatically populates the value "0035" if the classroom value is null.